moved image import files to lib/ started refactoring resume page to use new section
