+++ src/styles/resume.module.css
@@ -1,0 +1,3 @@
+.resume {
+
+}
